requirements.txt
streamlit
openai
jinja2
python-dotenv
litellm